% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_sum_matrix.R
\name{block_sum_matrix}
\alias{block_sum_matrix}
\title{Downsample a matrix by summing blocks of values}
\usage{
block_sum_matrix(
  matrix,
  downsample_factor_rows,
  downsample_factor_cols = downsample_factor_rows,
  average = F
)
}
\arguments{
\item{matrix}{The input matrix to downsample}

\item{downsample_factor_rows}{The downsampling factor for rows}

\item{downsample_factor_cols}{The downsampling factor for columns (default: \code{downsample_factor_rows})}

\item{average}{perform average instead of sum}
}
\value{
A downsampled matrix
}
\description{
Downsample a matrix by summing blocks of values
}
\details{
This function downsample a matrix by summing blocks of values. It takes an input matrix and two downsampling factors for rows and columns. It then creates downsampling matrices for rows and columns based on the provided factors. It then performs the downsampling operation by multiplying the input matrix with the transpose of the row downsampling matrix, then with the input matrix again, and finally with the column downsampling matrix. The result is a downsampled matrix.
}
\examples{
matrix(1, nrow = 5, ncol = 4) \%>\% block_sum_matrix(5, 2)

}
