% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_glmm_nb_hurdle.R
\name{bootstrap_glmm_nb_hurdle}
\alias{bootstrap_glmm_nb_hurdle}
\title{Perform Bootstrapped GLMM with Negative Binomial Hurdle Model}
\usage{
bootstrap_glmm_nb_hurdle(
  cell_name,
  formula_text = "value ~ treatment",
  long_ctrl_trt_df,
  fam = glmmTMB::truncated_nbinom2(),
  n_straps = 1000,
  n.cores = parallel::detectCores()/4,
  strap_size = NULL,
  return_full_results = F
)
}
\arguments{
\item{cell_name}{Character. The name of the cell for which the bootstrap analysis is to be performed. The dataset is filtered based on this value.}

\item{formula_text}{Character. A formula in text format specifying the model to be fitted. Default is `"value ~ treatment"`.}

\item{long_ctrl_trt_df}{DataFrame. A long-format data frame containing the control and treatment data.}

\item{fam}{Family. The family to be used in the GLMM model. Default is `glmmTMB::truncated_nbinom2()`.}

\item{n_straps}{Numeric. The number of bootstrap iterations. Default is `1000`.}

\item{n.cores}{Numeric. The number of cores to be used for parallel processing. Default is `parallel::detectCores()/4`.}

\item{strap_size}{Numeric. The size of the bootstrap samples. If `NULL`, the sample size is set to the number of rows in the filtered data. Default is `NULL`.}

\item{return_full_results}{Logical. If `TRUE`, returns both the full bootstrap results and the summarized results. If `FALSE`, only the summarized results are returned. Default is `FALSE`.}
}
\value{
A data frame containing the summarized bootstrap results, including estimates, standard errors, AIC values, and confidence intervals. If `return_full_results` is `TRUE`, a list is returned with both the full bootstrap results and the summarized results.
}
\description{
This function performs a bootstrapped Generalized Linear Mixed Model (GLMM) using a truncated negative binomial family. 
The function applies the model to a specified cell's data, fitting the model multiple times with resampled data 
to estimate the variability of the model's parameters.
}
\examples{
\dontrun{
  result <- bootstrap_glmm_nb_hurdle(
    cell_name = "cell_1",
    formula_text = "value ~ treatment",
    long_ctrl_trt_df = my_data_frame,
    n_straps = 500,
    n.cores = 4,
    strap_size = 1000,
    return_full_results = TRUE
  )
}

}
