% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_thresh.R
\name{cap_thresh}
\alias{cap_thresh}
\title{Cap and Threshold Matrix Values}
\usage{
cap_thresh(mat, cap = 20, thresh = 5, quantile = 0, zero_top = F)
}
\arguments{
\item{mat}{matrix to be capped and thresholded}

\item{cap}{the maximum value to be allowed in the matrix}

\item{thresh}{the minimum value to be allowed in the matrix}

\item{quantile}{the quantile to be used as the cap value}

\item{zero_top}{logical, indicating whether to zero out top values
(if 0, then the cap value is used)}
}
\value{
A cleaned matrix, after applying the cap and threshold
}
\description{
Eliminates outliers from the matrix, in both directions
}
\examples{
cap_thresh(matrix(1:100,10,10),cap=20,thresh=5)
}
\keyword{cap}
\keyword{matrix}
\keyword{threshold}
