% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_normalized_counts.R
\name{convert_to_normalized_counts}
\alias{convert_to_normalized_counts}
\title{Convert to Normalized Counts}
\usage{
convert_to_normalized_counts(dgel)
}
\arguments{
\item{dgel}{A data frame or list containing the fields `samples`, `norm.factors`, `lib.size`, and `counts`.
`samples` should be a list or data frame that includes `norm.factors` and `lib.size` for each sample.
`counts` should be a matrix or data frame of raw counts.}
}
\value{
A matrix of normalized counts.
}
\description{
This function adjusts the counts in a dataset by normalizing them based on the effective library size (ELS) and relative ELS.
It aims to reduce the impact of samples with high library size to give a fair comparison between condtions with different library sizes.
}
\examples{
# Assuming `dgel` is your dataset with the appropriate structure:
normalized_counts <- convert_to_normalized_counts(dgel)
}
\references{
For more information on the normalization process and its rationale, see the discussion at:
https://support.bioconductor.org/p/132240/
}
