% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_mappability_gc_bias.R
\name{correct_mappability_gc_bias}
\alias{correct_mappability_gc_bias}
\title{Correct Mappability and GC Bias in Count Matrices}
\usage{
correct_mappability_gc_bias(original_mat, dt_with_map_and_gc)
}
\arguments{
\item{original_mat}{A matrix of original counts to be corrected. Each row and column should represent different genomic features or samples.}

\item{dt_with_map_and_gc}{A data table containing mappability and GC content information for genomic regions. The table should have columns for chromosome (`chr`), start and end positions (`start`, `end`), mappability (`map_24bp`), and GC content (`gc_1bp`).}
}
\value{
A matrix of counts corrected for mappability and GC content biases, with the same dimensions and row/column names as the input matrix.
}
\description{
Use a negative binomial generalized linear model to adjust counts based on mappability and GC content.
}
\examples{
# Assuming `original_mat` is your matrix of counts and `dt_with_map_and_gc` is your data table with mappability and GC content:
corrected_mat <- correct_mappability_gc_bias(original_mat, dt_with_map_and_gc)

}
\references{
Hu M, Deng K, Selvaraj S, Qin ZS, Ren B and Liu JS (2012) HiCNorm: removing biases in Hi-C data via Poisson regression. (2012) Bioinformatics 28 (23), 3131-3133.
}
