% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bias_matrix.R
\name{create_bias_matrix}
\alias{create_bias_matrix}
\title{Create a MNase Bias Matrix}
\usage{
create_bias_matrix(
  chr = "chr15",
  start = 61982445,
  end = 61988445,
  genome = BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10,
  map_fn = system.file(package = "basepairC", "data/mm10/k24.umap.bedgraph.gz"),
  nrow = 6000,
  ncol = 6000
)
}
\arguments{
\item{chr}{A character string specifying the chromosome of the region of interest. Default is `"chr15"`.}

\item{start}{An integer specifying the start position of the region of interest. Default is `61982445`.}

\item{end}{An integer specifying the end position of the region of interest. Default is `61988445`.}

\item{genome}{A BSgenome object representing the genome of interest. Default is `BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10`.}

\item{map_fn}{A character string specifying the file path to the map file or `"mm10_k24"` to download it automatically. Default is the `k24.umap.bedgraph.gz` file from the `basepairC` package.}

\item{nrow}{An integer specifying the number of rows in the output matrix. Default is `6000`.}

\item{ncol}{An integer specifying the number of columns in the output matrix. Default is `6000`.}
}
\value{
A numeric matrix of size `nrow` x `ncol`, where each element represents the MNase-seq bias for a given genomic position.
}
\description{
This function generates a bias matrix based on the MNase-seq bias profile for a given genomic region.
}
\details{
Using the exact genomic sequence of the specified region,
constructs a bias matrix by multiplying a bias row vector with its transpose,
 resulting in a symmetric matrix representing the MNase-seq bias.
}
\examples{
\dontrun{
bias_matrix <- create_bias_matrix(chr="chr15", start=61982445, end=61988445)
}

}
