% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_long_ctrl_trt_df.R
\name{create_long_ctrl_trt_df}
\alias{create_long_ctrl_trt_df}
\title{Create Long-Form Control and Treatment Data Frame}
\usage{
create_long_ctrl_trt_df(
  ctl_filename,
  trt_filename,
  bin_size = 3,
  thresh = 5,
  cap = 20,
  lambda = 5,
  eps = 1e-10,
  maxIter = 20,
  quantile = 0,
  zero_top = T,
  chr = "chr15",
  start_row = 61982445,
  end_row = 61988445,
  start_col = 61982445,
  end_col = 61988445,
  dsfactor = 1,
  mat_start_row = 1,
  mat_end_row = 2000,
  mat_start_col = 1,
  mat_end_col = 2000,
  plot = F,
  ctrl_seg_max = 9,
  trt_seg_max = 7,
  lasso_ratio_seg_max = 7,
  ctrl_seg_method = "G",
  trt_seg_method = "G",
  lasso_ratio_seg_method = "G",
  ctrl_seg_model = "Dplus",
  trt_seg_model = "Dplus",
  lasso_ratio_seg_model = "Dplus",
  debug = F,
  n_cores = 1,
  make_symmetric = T,
  ds_fn = "sum",
  seg_algorithm = "HiCseg",
  ctrl_lasso = T,
  trt_lasso = T,
  total_read_correction = F,
  parallel_message = T,
  comp_type = "ratio",
  smooth = T,
  messages = F,
  manual_lr_breakpoints = "",
  ...
)
}
\arguments{
\item{ctl_filename}{The filename of the control matrix.}

\item{trt_filename}{The filename of the treatment matrix.}

\item{bin_size}{The bin size for genomic binning (default is 3).}

\item{thresh}{The threshold value for data processing (default is 5).}

\item{cap}{The cap value for data processing (default is 20).}

\item{lambda}{The regularization parameter for lasso filtering (default is 5).}

\item{eps}{The convergence threshold for lasso filtering (default is 1e-10).}

\item{maxIter}{The maximum number of iterations for lasso filtering (default is 1e4).}

\item{quantile}{The quantile value for data processing (default is 0.99).}

\item{zero_top}{Logical, indicating whether to zero out top values (default is TRUE).}

\item{chr}{The chromosome name (default is "chr15").}

\item{start_row}{The start position of the genomic region (default is 61982445).}

\item{end_row}{The end position of the genomic region (default is 61988445).}

\item{start_col}{The start position of the genomic region (default is 61982445).}

\item{end_col}{The end position of the genomic region (default is 61988445).}

\item{dsfactor}{The downsampling factor (default is 1).}

\item{plot}{Logical, indicating whether to generate plots (default is FALSE).}

\item{ctrl_seg_max}{The maximum number of segments for control (default is 9).}

\item{trt_seg_max}{The maximum number of segments for treatment (default is 7).}

\item{lasso_ratio_seg_max}{The maximum number of segments for lasso ratio (default is 7).}

\item{ctrl_seg_method}{The segmentation method for control (default is "B").}

\item{trt_seg_method}{The segmentation method for treatment (default is "B").}

\item{lasso_ratio_seg_method}{The segmentation method for lasso ratio (default is "G").}

\item{ctrl_seg_model}{The segmentation model for control (default is "Dplus").}

\item{trt_seg_model}{The segmentation model for treatment (default is "Dplus").}

\item{lasso_ratio_seg_model}{The segmentation model for lasso ratio (default is "Dplus").}

\item{debug}{Logical, indicating whether to enter debug mode (default is FALSE).}

\item{make_symmetric}{Logical, indicating whether to make the matrix symmetric (default is TRUE).}

\item{seg_algorithm}{The segmentation algorithm to use (default is "HiCseg"). jointSeg also works}

\item{ctrl_lasso}{Perform 2D LASSO on ctrl?}

\item{trt_lasso}{Perform 2D LASSO on trt?}

\item{total_read_correction}{Perform total read correction?}

\item{parallel_message}{Logical, indicating whether to display parallel message (default is TRUE).}

\item{mat_start}{The start index for matrix subsetting (default is 1).}

\item{mat_end}{The end index for matrix subsetting (default is 2000).}
}
\value{
A processed and segmented data frame.
}
\description{
This function reads, preprocesses, and segments control and treatment matrices
to create a long-form data frame. It performs various data processing steps
and segmentation to prepare the data for further analysis.
}
\examples{
ctl_file <- "control_matrix.csv"
trt_file <- "treatment_matrix.csv"
df <- create_long_ctrl_trt_df(ctl_file, trt_file)

}
