% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mat_list.R
\name{create_mat_list}
\alias{create_mat_list}
\title{Create a List of Matrices from rr_output}
\usage{
create_mat_list(rr_output, ncol_output = NULL, nrow_output = NULL, ncores = 1)
}
\arguments{
\item{rr_output}{A list containing at least a `sample_mat` matrix, with samples 
in columns and observations in rows. Additionally, `underscored_positions_col` 
and `underscored_positions_row` should contain the column and row names, respectively.}

\item{ncol_output}{Optional. The number of columns for the output matrices. If 
not provided, the number of columns will be set to the square root of the 
number of rows in `rr_output$sample_mat`.}

\item{nrow_output}{Optional. The number of rows for the output matrices. If 
not provided, the number of rows will be set to the square root of the 
number of rows in `rr_output$sample_mat`.}

\item{ncores}{Optional. The number of cores to use for parallel processing. 
If set to "auto", the number of cores will be automatically detected based 
on the available cores on the machine. Default is 1.}
}
\value{
A list of matrices. Each matrix corresponds to one sample from 
  `rr_output$sample_mat` and has dimensions `nrow_output` by `ncol_output`. 
  The matrices will also have the attributes `condition` and `sample_name`.
}
\description{
This function processes the output of `rr_output` and returns a list of matrices,
one for each sample in the input matrix. Each matrix will have specified row and 
column names and attributes, including `condition` and `sample_name`.
}
\details{
Creates a list of matrices, with condition and sample name attribute information.
   The column and row names of each matrix are taken from 
  `rr_output$underscored_positions_col` and `rr_output$underscored_positions_row`, 
  respectively. Each matrix is stored with attributes for `condition` and 
  `sample_name`.
}
\examples{
\dontrun{
rr_output <- list(
  sample_mat = matrix(rnorm(100), 10, 10), 
  underscored_positions_col = paste0("col_", 1:10), 
  underscored_positions_row = paste0("row_", 1:10)
)
create_mat_list(rr_output, ncores = "auto")
}
}
