% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parquet_set.R
\name{create_parquet_set}
\alias{create_parquet_set}
\title{Create a Parquet Dataset from a List of Matrices}
\usage{
create_parquet_set(
  mat_list,
  segments,
  bias_dt_melted,
  parquet_dir,
  final_parquet_file,
  low_disk_footprint = F,
  ncores = 1,
  mc.cores = 1,
  verbose = F
)
}
\arguments{
\item{mat_list}{A list of matrices to be processed, where each matrix is associated with a sample.}

\item{segments}{A segmentation object used for preprocessing the matrices.}

\item{bias_dt_melted}{A melted data.table containing bias matrices, to be merged with the processed matrices.}

\item{parquet_dir}{A directory path where the Parquet files will be written. The directory will be created if it does not exist.}

\item{final_parquet_file}{Path to the final Parquet dataset that will be created after processing all matrices.}

\item{low_disk_footprint}{Logical flag indicating whether to use a low disk footprint approach (default is \code{FALSE}).}

\item{ncores}{Number of cores to use for parallel processing (default is 1).}
}
\value{
\code{NULL}. The processed matrices are written to Parquet files in the specified directory.
}
\description{
This function processes a list of matrices, applies preprocessing steps, merges with a melted bias data table, and writes the resulting data to a Parquet dataset. 
It supports parallel processing and low disk footprint options.
Only choose the low disk footprint option if you have limited disk space and have enough memory to store the matrices (typically 512GB would be enough to comfortably do this).
}
\details{
This function applies the following steps to each matrix in \code{mat_list}:
\enumerate{
  \item Preprocesses the matrix using the provided segmentation object.
  \item Converts the preprocessed matrix to a data.table.
  \item Adds sample metadata to the data.table.
  \item Removes rows with specific treatment conditions.
  \item Extracts chromosome, start, and end positions from locational information.
  \item Renames the treatment column to distinguish between control and treatment groups.
  \item Merges the processed matrix with the bias data.table based on genomic coordinates.
  \item Writes the merged data to a Parquet file in the specified directory.
}
The function writes individual Parquet files for each matrix and then combines them into a Parquet dataset.
}
\examples{
\dontrun{
create_parquet_set(mat_list, segments, bias_dt_melted, "output/parquet_dir", "final_parquet_file", low_disk_footprint = TRUE, ncores = 4)
}

}
