% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ratio_mat.R
\name{create_ratio_mat}
\alias{create_ratio_mat}
\title{Create Ratio Matrix Function}
\usage{
create_ratio_mat(
  mat1,
  mat2,
  cap = 100,
  thresh,
  zero_top = F,
  quantile = 0,
  adjust_for_total_reads = "after_cleaning"
)
}
\arguments{
\item{mat1}{A numeric matrix.}

\item{mat2}{A numeric matrix.}

\item{cap}{The cap value for thresholding (default is 100).}

\item{thresh}{The threshold value for thresholding.}

\item{zero_top}{Logical, indicating whether to zero out top values (default is FALSE).}

\item{quantile}{The quantile value for thresholding (default is 0).}

\item{adjust_for_total_reads}{A character string, specifying whether to adjust for total reads "before_cleaning" or "after_cleaning".}
}
\value{
A numeric matrix with ratio values.
}
\description{
This function takes two matrices, performs various operations on them, and
returns a ratio matrix. It includes options for capping values, applying a
threshold, zeroing the top values, and adjusting for total reads.
}
\examples{
mat1 <- matrix(1:10, nrow = 2)
mat2 <- matrix(11:20, nrow = 2)
create_ratio_mat(mat1, mat2)

}
