% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_sample_mat.R
\name{downsample_col_multinomial}
\alias{downsample_col_multinomial}
\title{Downsample a column of read counts.}
\usage{
downsample_col_multinomial(col, min_total)
}
\arguments{
\item{col}{A numeric vector or matrix column representing the read counts to be downsampled.}

\item{prop_to_retain}{A numeric value or vector representing the proportion of reads to retain for each element of \code{col}. Should be between 0 and 1.}
}
\value{
A matrix with the same dimensions as \code{col} but with downsampled read counts.
}
\description{
This function downsamples a column of read counts by retaining a specified proportion
of the original counts. The downsampling is done using a binomial distribution.
}
