% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_sample_mat.R
\name{downsample_dgel}
\alias{downsample_dgel}
\title{Downsample counts in a DGEList object.}
\usage{
downsample_dgel(dgel)
}
\arguments{
\item{dgel}{A DGEList object from the \code{edgeR} package, containing the read counts to be downsampled.}
}
\value{
A DGEList object with downsampled counts.
}
\description{
This function downsamples the counts in a DGEList object to match the sample with the lowest total read count. The function works on the \code{counts} element of the DGEList.
}
