% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_sample_mat.R
\name{downsample_reads_by_sample}
\alias{downsample_reads_by_sample}
\title{Downsample read counts by sample.}
\usage{
downsample_reads_by_sample(sample_mat, type = "by_sample")
}
\arguments{
\item{sample_mat}{A matrix or data frame where each column represents a sample and each row represents a feature. The elements are read counts.}

\item{type}{A character string indicating the downsampling method. Currently, only "by_sample" is supported, which downsamples each sample to the minimum total read count across all samples.}
}
\value{
A matrix with downsampled read counts. The dimensions and names are the same as the input \code{sample_mat}.
}
\description{
This function downsamples the read counts in a sample matrix to match the sample with the lowest total read count. The downsampling is performed on a per-sample basis.
}
