% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_sample_mat.R
\name{downsample_rr}
\alias{downsample_rr}
\title{Downsample counts from the output of a read_replicates call.}
\usage{
downsample_rr(rrobj)
}
\arguments{
\item{dgel}{read_replicates output list}
}
\value{
A read_replicates output list with downsampled counts.
}
\description{
This function downsamples the counts in a read_replicates object to match the sample with the lowest total read count. The function works on the \code{counts} element of the read_replicates output.
}
