% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_smooth_mat_set.R
\name{ds_smooth_mat_set}
\alias{ds_smooth_mat_set}
\title{Smooth and Downsample Hi-C Matrices with MNase Bias Correction}
\usage{
ds_smooth_mat_set(
  ctrl_matrix,
  trt_matrix,
  mnase_bias_matrix,
  smooth_window = 1,
  ds_factor = 10,
  filter_action = c("mean", "max"),
  mean_ds = F
)
}
\arguments{
\item{ctrl_matrix}{A numeric matrix representing the control interaction counts.}

\item{trt_matrix}{A numeric matrix representing the treatment interaction counts.}

\item{mnase_bias_matrix}{A numeric matrix representing the MNase-seq bias, corresponding to the control and treatment matrices.}

\item{smooth_window}{An integer specifying the window size for smoothing the MNase bias matrix using a kernel smoothing function. Default is `100`.}

\item{ds_factor}{An integer specifying the downsampling factor to reduce the resolution of the matrices. Default is `10`.}

\item{filter_action}{A character vector specifying the type of filter to apply to the MNase bias matrix. Options include `mean` (default, using a kernel smooth) and `max` (using dilation).}
}
\value{
A list containing three matrices:
\item{ctrl}{The smoothed and downsampled control matrix.}
\item{trt}{The smoothed and downsampled treatment matrix.}
\item{mnase_bias}{The smoothed and downsampled MNase bias matrix.}
}
\description{
This function smooths and downsamples control and treatment matrices along with a MNase bias matrix.
}
\details{
The function first sets the row and column names of the MNase bias matrix to match those of the control matrix. It then performs a kernel smoothing operation on the MNase bias matrix to account for proximal effects of MNase-seq bias. After smoothing, the function downsamples the control, treatment, and MNase bias matrices using the specified downsampling factor.
}
\examples{
\dontrun{
smoothed_downsampled_data <- ds_smooth_mat_set(ctrl_matrix, trt_matrix, mnase_bias_matrix)
ctrl_matrix_smoothed <- smoothed_downsampled_data$ctrl
trt_matrix_smoothed <- smoothed_downsampled_data$trt
mnase_bias_smoothed <- smoothed_downsampled_data$mnase_bias
}

}
