% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encompass.R
\name{encompass}
\alias{encompass}
\title{Get Encompassing Region of a Genomic Ranges Object}
\usage{
encompass(gr)
}
\arguments{
\item{gr}{A GRanges object}
}
\value{
A single GRange, encompassing the max and min of the input GRanges object
}
\description{
Get Encompassing Region of a Genomic Ranges Object
}
\keyword{genomicranges}
\keyword{iranges}
