% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalize_sample_coverage.R
\name{equalize_sample_coverage}
\alias{equalize_sample_coverage}
\title{Equalize Sample Coverage}
\usage{
equalize_sample_coverage(
  sample_mat,
  n_samples,
  max_cores = parallel::detectCores()
)
}
\arguments{
\item{sample_mat}{A data frame or matrix where columns represent samples. Columns can be a mix of numeric and non-numeric data.}

\item{n_samples}{An integer indicating the number of samples to generate in the equalization process.}
}
\value{
A data frame with the same dimensions as `sample_mat`, where numeric columns have had their coverage equalized
  across samples, and non-numeric columns are returned unchanged.
}
\description{
This function takes a matrix of samples and equalizes the coverage by sampling non-zero entries.
It handles numeric and non-numeric columns separately, ensuring that the numeric columns have equalized
coverage across samples. The non-numeric columns are returned unchanged.
}
\details{
The function first separates numeric and non-numeric columns. It then processes the numeric columns to equalize coverage 
by sampling non-zero entries in each column such that the number of non-zero entries matches the minimum across all columns.
Finally, it recombines the numeric and non-numeric columns and returns the result.
}
\examples{
\dontrun{
# Example usage:
sample_matrix <- data.frame(
  sample1 = c(1, 2, 0, 4),
  sample2 = c(0, 5, 6, 0),
  sample3 = c(7, 0, 9, 10),
  non_numeric = c("A", "B", "C", "D")
)
n_samples <- 100
equalized_samples <- equalize_sample_coverage(sample_matrix, n_samples)
}

}
