% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_lasso_filter.R
\name{fused_lasso_filter}
\alias{fused_lasso_filter}
\title{Fused Lasso Filtering Function}
\usage{
fused_lasso_filter(mat, lambda, eps, maxIter)
}
\arguments{
\item{mat}{A numeric matrix to be filtered.}

\item{lambda}{A numeric value representing the regularization parameter.}

\item{eps}{Convergence threshold paramter (see glmgen::fusedLattice).}

\item{maxIter}{maximum number of iterations for convergence.}
}
\value{
A filtered numeric matrix.
}
\description{
This function applies the fused lasso filtering technique to a given matrix.
}
\examples{
mat <- matrix(1:10, nrow = 2)
fused_lasso_filter(mat, lambda = 0.1, eps = 0.001, maxIter = 100)

}
