% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam2d_negbin_surface_model.R
\name{gam2d_negbin_surface_model}
\alias{gam2d_negbin_surface_model}
\title{gam2d_negbin_surface_model}
\usage{
gam2d_negbin_surface_model(long_ctrl_trt_df, knots = 100, plot = T)
}
\arguments{
\item{long_ctrl_trt_df}{A long-format data frame with control/treatment ratio data.}

\item{knots}{The number of knots to use in the 2D GAM model.}

\item{plot}{A logical indicating whether to plot the 2D GAM model.}
}
\value{
pred_gam2d A data frame with the predicted values and standard errors from the 2D GAM model.
}
\description{
This function fits a 2D GAM model to the control/trt ratios using a negative binomial distribution.
}
