% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_win.R
\name{gc_win}
\alias{gc_win}
\title{Calculate GC Content in a Sliding Window}
\usage{
gc_win(
  chromosome,
  position,
  genome = BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10,
  window_size = 5,
  n_cores = parallel::detectCores()/2
)
}
\arguments{
\item{chromosome}{The chromosome for which to calculate GC content.}

\item{position}{A vector of positions on the chromosome to calculate GC content for.}

\item{genome}{The genome object from which to retrieve sequence data. By default, this is set to the mouse genome (BSgenome.Mmusculus.UCSC.mm10).}

\item{window_size}{The size of the sliding window for GC content calculation. Default is 5.}

\item{n_cores}{The number of cores to use for parallel processing. Default is half of the available cores detected on the system.}
}
\value{
A vector of GC content values for each position, calculated within the specified sliding window.
}
\description{
This function calculates the GC content within a sliding window across specified genomic positions on a chromosome. It is designed to work with genomic data, utilizing a specified genome object for sequence retrieval.
}
\examples{
# Calculate GC content for positions 100 to 200 on chromosome 1
gc_content <- gc_win(
  chromosome = "chr1",
  position = 100:200,
  genome = BSgenome.Mmusculus.UCSC.mm10,
  window_size = 5
)

}
