% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{generate_distibution_plots}
\alias{generate_distibution_plots}
\title{Generate distribution plots}
\usage{
generate_distibution_plots(df, treatment)
}
\arguments{
\item{df}{A long-format data frame with control and treatment ratio data.}

\item{treatment}{A character string indicating the treatment group.}
}
\value{
output_list A list of distribution plots and two data frames with distribution statistics for before the transformation and after the transformation.
}
\description{
This function generates distribution plots for ctrl/trt ratios and their transformations.
}
