% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_area.R
\name{get_cell_area}
\alias{get_cell_area}
\title{Calculate Area of a Cell Based on ID}
\usage{
get_cell_area(cell_id)
}
\arguments{
\item{cell_id}{A character string representing the cell ID, which includes two position
data points separated by a space. Each position should be in a format
compatible with `CNVScope::underscored_pos_to_GRanges`.}
}
\value{
Returns a numeric value representing the area of the cell.
}
\description{
This function calculates the area of a cell given a cell ID that encodes position data.
The cell ID should contain two position data points separated by a space. Each position data
is converted to a GRanges object using `CNVScope::underscored_pos_to_GRanges`, from which
the width is extracted. The area is computed as the product of the widths of these two
positions.
}
\examples{
# Example cell ID format: "chr1_100_200 chr2_150_250"
get_cell_area("chr1_100_200 chr2_150_250") # Calculates the area based on positions.

}
