% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditions.R
\name{get_conditions}
\alias{get_conditions}
\title{Get Conditions for Replicate Files}
\usage{
get_conditions(replicate_files, control_tag = "WT")
}
\arguments{
\item{replicate_files}{A character vector of file names or paths representing the replicate files.}

\item{control_tag}{A character string representing the tag for control files. Defaults to `"WT"`.}
}
\value{
A list with two elements:
\item{conditions}{A character vector representing the conditions for each file ("ctrl" or "trt").}
\item{replicate_files}{A character vector of replicate file names, ordered by condition.}
}
\description{
This function assigns conditions ("ctrl" for control and "trt" for treatment) to replicate files based on a specified control tag. It returns a list containing the conditions and the ordered replicate file names.
The condtiions are ordered by the condition.
}
\examples{
replicate_files <- c("WT_sample1.bam", "TRT_sample2.bam", "WT_sample3.bam")
result <- get_conditions(replicate_files, control_tag = "WT")
print(result)

}
