% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctcf_segments.R
\name{get_ctcf_segments}
\alias{get_ctcf_segments}
\title{Get CTCF Segments}
\usage{
get_ctcf_segments(
  chr = "chr15",
  start = 61982445,
  end = 61988445,
  genome_name = "mm10",
  sample_mat_list = NULL,
  return_gr = F,
  overlap_gap = 0
)
}
\arguments{
\item{chr}{A character string indicating the chromosome (default is "chr15").}

\item{start}{An integer representing the start position of the genomic region (default is 61982445).}

\item{end}{An integer representing the end position of the genomic region (default is 61988445).}

\item{genome_name}{A character string indicating the genome version (default is "mm10").}

\item{sample_mat_list}{A list of matrices representing sample data.}
}
\value{
A list containing breakpoint positions for rows and columns, and their transposed versions.
}
\description{
This function retrieves CTCF binding sites from a genomic region, segments a sample matrix list,
and returns the corresponding breakpoints.
}
