% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff_segments.R
\name{get_diff_segments}
\alias{get_diff_segments}
\title{Get Segments For a Two-Condition Comparison}
\usage{
get_diff_segments(
  sample_mat_list,
  conditions = NULL,
  comp_type = c("absdiff", "loss", "gain", "ratio"),
  return_df = F,
  lambda = NULL,
  upperQ = 0.05,
  lowerQ = 0.95
)
}
\arguments{
\item{sample_mat_list}{A list of matrices representing sample data.}

\item{conditions}{A character vector indicating conditions (e.g., "ctrl" or "trt"). It must have exactly two unique values.}

\item{comp_type}{A character string indicating the type of comparison (options are "absdiff", "loss", "gain", or "ratio").}
}
\value{
A list containing segmented breakpoints for rows and columns, and their transposed versions.
}
\description{
This function computes the difference between two conditions and returns segmented breakpoints based on the difference.
}
