% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_segments.R
\name{get_grid_segments}
\alias{get_grid_segments}
\title{Get grid segments from a sample matrix list}
\usage{
get_grid_segments(sample_mat_list, every_bp = 500)
}
\arguments{
\item{sample_mat_list}{A list of matrices. The function assumes all matrices have the same dimensions, and it uses the first matrix in the list to determine the dimensions.}

\item{every_bp}{An integer specifying the interval (in base pairs) at which to define breakpoints for the matrix grid. Default is 500.}
}
\value{
A list containing four elements:
\describe{
  \item{breakpoints_col}{A vector of breakpoints for the columns of the matrix.}
  \item{breakpoints_row}{A vector of breakpoints for the rows of the matrix.}
  \item{t_breakpoints_col}{A vector of breakpoints for the columns of the transposed matrix.}
  \item{t_breakpoints_row}{A vector of breakpoints for the rows of the transposed matrix}
}
}
\description{
This function calculates grid breakpoints for the rows and columns of a matrix, based on a specified interval (`every_bp`). 
The breakpoints represent the positions at which segments of the matrix are defined. The function returns a list with 
breakpoints for rows, columns, and their transposed equivalents.
}
\examples{
# Example usage:
mat1 <- matrix(1:10000, nrow = 100, ncol = 100)
mat_list <- list(mat1)
seg_obj <- get_grid_segments(mat_list, every_bp = 20)
print(seg_obj)

}
