% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_segments.R
\name{get_sum_segments}
\alias{get_sum_segments}
\title{Get Sum of Segments from Sample Matrix List}
\usage{
get_sum_segments(sample_mat_list, n_segs = 9)
}
\arguments{
\item{sample_mat_list}{A list of matrices, where each matrix corresponds to a sample.}

\item{n_segs}{An integer representing the number of segments for both rows and columns in the Lasso segmentation (default is 9).}
}
\value{
A list containing:
\describe{
  \item{lasso_segments}{Lasso-based segmentation results.}
  \item{orig_segments}{Original segmentation results.}
}
}
\description{
This function computes the sum of matrices from a list of sample matrices and processes
the resulting matrix by applying a segmentation preprocessing step.
}
\examples{
\dontrun{
  get_sum_segments(sample_mat_list, n_segs = 9)
}
}
