% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_compare.R
\name{glm_compare}
\alias{glm_compare}
\title{parallel_compare}
\usage{
glm_compare(long_ctrl_trt_df, interactive = F)
}
\arguments{
\item{long_ctrl_trt_df}{A long-format data frame with control and treatment ratio data.}

\item{interactive}{A logical value indicating whether to plot the results interactively (default is FALSE).}
}
\value{
A data frame with the treatment effect comparison results.
}
\description{
This function performs a parallel comparison of the treatment effect of a control and treatment group.
}
\examples{
dontrun{
glm_compare(long_ctrl_trt_df, interactive = TRUE)
}
}
