% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iset_to_gint.R
\name{iset_to_gint}
\alias{iset_to_gint}
\title{Convert InteractionSet to GenomicInteractions Object}
\usage{
iset_to_gint(iset)
}
\arguments{
\item{iset}{An `InteractionSet` object to be converted.}
}
\value{
A `GenomicInteractions` object containing the interactions defined in the `iset` object.
}
\description{
This function takes an `InteractionSet` object and converts it into a `GenomicInteractions` object. It extracts the first and second anchors from the `InteractionSet` to create the `GenomicInteractions` object.
}
\examples{
# Assuming `iset` is an InteractionSet object
gint <- iset_to_gint(iset)

}
