% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso2d.R
\name{lasso2d}
\alias{lasso2d}
\title{Perform 2D LASSO regularization on a matrix}
\usage{
lasso2d(mat, lambda = 5, eps = 0.01, maxIter = 20, remove_zero = F)
}
\arguments{
\item{mat}{A numeric matrix on which 2D LASSO regularization is to be applied.}

\item{lambda}{Regularization parameter for LASSO. Defaults to 5.}

\item{eps}{Convergence threshold for the algorithm. Defaults to 1e-2.}

\item{maxIter}{Maximum number of iterations for the algorithm. Defaults to 20.}

\item{remove_zero}{Logical indicating whether to remove zero elements from the regularized matrix. Defaults to FALSE.}
}
\value{
A list containing the regularized matrix (`lasso_mat`) and the beta matrix (`beta_mat`). If `remove_zero` is TRUE, `lasso_mat` will only contain non-zero elements.
}
\description{
This function applies 2D LASSO regularization on a given matrix using the fusedLattice function from the glmgen package. It returns the regularized matrix and the beta matrix used for regularization. Optionally, it can remove elements of the matrix that are zero after regularization.
}
\examples{
mat <- matrix(rnorm(100), 10, 10)
result <- lasso2d(mat)
str(result)

}
