% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_1d_filter.R
\name{lasso_1d_filter}
\alias{lasso_1d_filter}
\title{LASSO-Based Filtering of Data Frame for 1D Covariates}
\usage{
lasso_1d_filter(long_df, mc.cores = 1, return_df = F, thresh = -Inf, cap = Inf)
}
\arguments{
\item{long_df}{A data frame in long format. The data frame should contain at least 
the columns \code{treatment}, \code{trtlr_cell_id}, and \code{value}. The function
will filter for rows where \code{treatment} is either "ctrl" or "trt", and \code{value}
falls between 10 and a specified cap value (available in the global environment as \code{cap_val}).}

\item{mc.cores}{The number of cores to use for parallel computation during cross-validation in LASSO.
Default is 1.}

\item{return_df}{Logical; if \code{TRUE}, the function returns a data frame of retained
rows. If \code{FALSE} (default), it returns a vector of \code{trtlr_cell_id}s that
were removed by the filtering process.}
}
\value{
If \code{return_df = TRUE}, a filtered version of the input data frame is returned.
  If \code{return_df = FALSE} (default), a vector of removed \code{trtlr_cell_id}s is returned.
}
\description{
Filters a long-format data frame based on LASSO regression, removing cell_ids based on a quasipoisson model.
}
\details{
The function first filters the input data frame based on specific conditions: it retains
only rows where \code{treatment} is "ctrl" or "trt", and where \code{value} is greater than 10
and less than a global \code{cap_val}. It renames the column \code{trtlr_cell_id} to \code{cell_id} 
for the LASSO analysis. A LASSO regression with cross-validation is then applied using the 
quasipoisson family. The retained covariates (non-zero beta coefficients) are identified, 
and either the filtered data frame or the list of removed covariates is returned, depending on 
the \code{return_df} parameter.
}
\examples{
\dontrun{
  filtered_df <- lasso_1d_filter(my_long_df, mc.cores = 4, return_df = TRUE)
  removed_ids <- lasso_1d_filter(my_long_df, mc.cores = 2, return_df = FALSE)
}

}
