% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_symmetric.R
\name{make_symmetric}
\alias{make_symmetric}
\title{Enforce matrix symmetry}
\usage{
make_symmetric(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A symmetric matrix
}
\description{
Adds the upper and lower triangular parts, after forcing symmetry on each
}
\keyword{lower}
\keyword{matrix}
\keyword{symmetric}
\keyword{symmetry}
\keyword{upper}
