% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_win.R
\name{map_win}
\alias{map_win}
\title{Calculate Mappability Content in a Sliding Window}
\usage{
map_win(
  chromosome = gene_locus_dt_bp$chr,
  position = gene_locus_dt_bp$start,
  map_gr = mappability_gr,
  window_size = 200,
  n_cores = parallel::detectCores()/2
)
}
\arguments{
\item{chromosome}{The chromosome for which to calculate mappability content. Defaults to the chromosome column in `gene_locus_dt_bp`.}

\item{position}{A vector of positions on the chromosome to calculate mappability content for. Defaults to the start column in `gene_locus_dt_bp`.}

\item{map_gr}{A `GenomicRanges` object containing mappability data for the genome.}

\item{window_size}{The size of the sliding window for mappability content calculation. Default is 200.}

\item{n_cores}{The number of cores to use for parallel processing. Default is half of the available cores detected on the system.}
}
\value{
A vector of mappability content values for each position, calculated within the specified sliding window.
}
\description{
This function calculates the mean mappability content within a sliding window for specified genomic positions on a chromosome with a `GenomicRanges` object that specifies mappability.
}
\examples{
# Assuming `gene_locus_dt_bp` and `mappability_gr` are defined
mappability_content <- map_win(
  chromosome = gene_locus_dt_bp$chr,
  position = gene_locus_dt_bp$start,
  map_gr = mappability_gr,
  window_size = 200
)

}
