% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pooling_parallel.R
\name{mean_pooling_parallel}
\alias{mean_pooling_parallel}
\title{Mean Pooling with Parallel Processing Function}
\usage{
mean_pooling_parallel(mat, factor, ncores = parallel::detectCores())
}
\arguments{
\item{mat}{A numeric matrix to be pooled.}

\item{factor}{The pooling factor for aggregating matrix elements.}

\item{ncores}{The number of CPU cores to use for parallel processing (default is the number of detected cores).}
}
\value{
A pooled numeric matrix using mean pooling.
}
\description{
This function performs mean pooling on a matrix with parallel processing support.
}
\examples{
mat <- matrix(1:16, nrow = 4)
mean_pooling_parallel(mat, factor = 2)

}
