% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimaxnorm.R
\name{minmaxnorm}
\alias{minmaxnorm}
\title{Min-Max Normalization Function}
\usage{
minmaxnorm(x)
}
\arguments{
\item{x}{A numeric vector to be normalized.}
}
\value{
A numeric vector with values normalized to the range [0, 1].
}
\description{
Min-Max Normalization Function
}
\examples{
x <- c(1, 2, 3, 4, 5)
minmaxnorm(x)
}
