% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnase_correct_matrix_pair.R
\name{mnase_correct_matrix_pair}
\alias{mnase_correct_matrix_pair}
\title{Correct MNase Bias in Hi-C Matrices for Control and Treatment}
\usage{
mnase_correct_matrix_pair(
  ctrl_mat_ds,
  trt_mat_ds,
  mnase_bias_matrix_sm_ds,
  snr_filter = F,
  two_stage = F,
  single_model_cores = 1,
  residual = F
)
}
\arguments{
\item{ctrl_mat_ds}{A numeric matrix representing the control Hi-C interaction counts.}

\item{trt_mat_ds}{A numeric matrix representing the treatment Hi-C interaction counts.}

\item{mnase_bias_matrix_sm_ds}{A numeric matrix representing the MNase-seq bias. This matrix should have the same dimensions as `ctrl_mat_ds` and `trt_mat_ds`.}

\item{snr_filter}{A logical value indicating whether to apply signal-to-noise ratio (SNR) filtering. Default is `FALSE`.}

\item{two_stage}{A logical value indicating whether to use a two-stage correction. Default is `FALSE`.}

\item{single_model_cores}{An integer specifying the number of cores to use for the single model correction. Default is `1`.}

\item{residual}{Does a correction with the residuals instead of outcome. Stronger SNR filter, though the residuals may have a longer tailed distribution. Default is `FALSE`.}
}
\value{
A list containing two corrected Hi-C matrices: 
\item{ctrl}{Corrected control Hi-C matrix.}
\item{trt}{Corrected treatment Hi-C matrix.}
}
\description{
This function corrects the MNase-seq bias in two Hi-C matrices (control and treatment) using a pre-computed MNase bias matrix.
}
\details{
The function first melts the MNase bias matrix and joins it with the melted control and treatment matrices. It then applies a Negative Binomial generalized linear model (GLM) to correct for MNase bias. The correction factor is calculated based on the GLM coefficients and applied to the original interaction counts to produce the corrected matrices. The function outputs the corrected matrices in a list.
}
\examples{
\dontrun{
corrected_matrices <- mnase_correct_matrix_pair(ctrl_mat_ds, trt_mat_ds, mnase_bias_matrix_sm_ds)
corrected_ctrl <- corrected_matrices$ctrl
corrected_trt <- corrected_matrices$trt
}

}
