% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_approx.R
\name{nmf_approx}
\alias{nmf_approx}
\title{Approximate a Matrix using Non-negative Matrix Factorization (NMF)}
\usage{
nmf_approx(mat, N, nmf_method = "lee")
}
\arguments{
\item{mat}{A numeric matrix. The input matrix to be approximated.}

\item{N}{An integer. The number of components to use for the NMF approximation.}

\item{nmf_method}{A character string. The method to use for the NMF algorithm. Default is "lee".}
}
\value{
A numeric matrix. The approximated matrix.
}
\description{
This function approximates a given matrix using Non-negative Matrix Factorization (NMF).
}
\details{
This function performs Non-negative Matrix Factorization (NMF) on the input matrix `mat`
using the specified number of components `N` and the specified method `nmf_method`. It returns the
approximated matrix obtained by multiplying the basis matrix `W` and the coefficient matrix `H`.
}
\examples{
\dontrun{
mat <- matrix(runif(100), nrow = 10)
N <- 3
approx_mat <- nmf_approx(mat, N)
}

}
