% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zminmaxnorm.R
\name{noninf}
\alias{noninf}
\title{Finite Elements Function}
\usage{
noninf(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with finite values.
}
\description{
This function takes a numeric vector and returns only the finite elements.
}
\examples{
x <- c(1, 2, Inf, 3, NaN)
noninf(x)

}
