% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonzero.R
\name{nonzero}
\alias{nonzero}
\title{Non-Zero Elements Function}
\usage{
nonzero(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with non-zero values.
}
\description{
This function takes a numeric vector and returns only the non-zero elements.
}
\examples{
x <- c(0, 1, 2, 0, 3)
nonzero(x)

}
