% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_compare.R
\name{parallel_compare}
\alias{parallel_compare}
\title{parallel_compare}
\usage{
parallel_compare(long_ctrl_trt_df, mc.cores = 1, interactive = F)
}
\arguments{
\item{long_ctrl_trt_df}{A long-format data frame with control and treatment ratio data.}

\item{mc.cores}{The number of CPU cores to use for parallel processing (default is 1).}

\item{interactive}{A logical value indicating whether to plot the results interactively (default is FALSE).}
}
\value{
A data frame with the treatment effect comparison results.
}
\description{
This function performs a parallel comparison of the treatment effect of a control and treatment group.
}
\examples{
dontrun{
parallel_compare(long_ctrl_trt_df, mc.cores = 2, interactive = TRUE) 
}
}
