% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_compare_np.R
\name{parallel_compare_np}
\alias{parallel_compare_np}
\title{parallel_compare (non-parametric)}
\usage{
parallel_compare_np(
  long_ctrl_trt_df,
  mc.cores = 1,
  interactive = F,
  alternative = "less"
)
}
\arguments{
\item{long_ctrl_trt_df}{A long-format data frame with control and treatment ratio data.}

\item{mc.cores}{The number of CPU cores to use for parallel processing (default is 1).}

\item{interactive}{A logical value indicating whether to plot the results interactively (default is FALSE).}
}
\value{
A data frame with the treatment effect comparison results.
}
\description{
This function performs a parallel comparison of the treatment effect of a control and treatment group with Wilcoxon & permutation tests.
}
\examples{
dontrun{
parallel_compare(long_ctrl_trt_df, mc.cores = 2, interactive = TRUE) 
}
}
