% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_mat.R
\name{preprocess_mat}
\alias{preprocess_mat}
\title{preprocess_mat}
\usage{
preprocess_mat(
  fn,
  cond,
  bin_size = 1,
  thresh = 0,
  cap = Inf,
  lambda = 5,
  eps = 1e-10,
  maxIter = 20,
  quantile = 0,
  zero_top = F,
  chr = "chr15",
  start_row = 61982445,
  end_row = 61988445,
  start_col = 61982445,
  end_col = 61988445,
  dsfactor = 50,
  mat_start_row = 1,
  mat_end_row = 6000,
  mat_start_col = 1,
  mat_end_col = 6000,
  plot = F,
  orig_seg_max_col = 9,
  orig_seg_max_row = 9,
  lasso_seg_max_row = 9,
  lasso_seg_max_col = 9,
  orig_seg_method = "G",
  lasso_seg_method = "G",
  orig_seg_model = "Dplus",
  lasso_seg_model = "Dplus",
  debug = F,
  n_cores = 1,
  make_symmetric = T,
  ds_fn = "sum",
  seg_algorithm = "jointSeg",
  hicseg_bp_selection = T,
  lasso2d = F,
  verbose = F,
  segments_orig_obj = NULL,
  segments_lasso_obj = NULL,
  seg_method = "RBS",
  smooth = T,
  smooth_window = 5,
  lasso_seg = F
)
}
\arguments{
\item{fn}{The file name of the matrix to be preprocessed.}

\item{cond}{The condition of the matrix.}

\item{bin_size}{The bin size of the matrix.}

\item{thresh}{Removes reads below this count value from the input matrix}

\item{cap}{Reads above this height will be discarded or thresholded.}

\item{lambda}{The lambda value for the fused 2D lasso.}

\item{eps}{The epsilon value for the fused 2D lasso.}

\item{maxIter}{The maximum number of iterations for the fused 2D lasso.}

\item{quantile}{The quantile value for the cap.}

\item{zero_top}{Logical, indicating whether to zero out top values (if 0, then the cap value is used).}

\item{chr}{The chromosome of the matrix.}

\item{start_row}{The starting row of the matrix.}

\item{end_row}{The ending row of the matrix.}

\item{start_col}{The starting column of the matrix.}

\item{end_col}{The ending column of the matrix.}

\item{dsfactor}{The downsample factor for the matrix (essentially a 6000x6000 matrix becomes a 2000x2000 matrix if the dsfactor is 3).}

\item{mat_start_row}{The starting row of the matrix.}

\item{mat_end_row}{The ending row of the matrix.}

\item{mat_start_col}{The starting column of the matrix.}

\item{mat_end_col}{The ending column of the matrix.}

\item{plot}{Logical, indicating whether to plot the matrix.}

\item{orig_seg_max_col}{The maximum column value for the original segmentation.}

\item{orig_seg_max_row}{The maximum row value for the original segmentation.}

\item{lasso_seg_max_row}{The maximum row value for the lasso segmentation.}

\item{lasso_seg_max_col}{The maximum column value for the lasso segmentation.}

\item{orig_seg_method}{The distribution for the original segmentation (P, G, or B for poisson, gaussian, or binomial, respectively).}

\item{lasso_seg_method}{The distribution for the lasso segmentation.}

\item{orig_seg_model}{The model for the original matrix segmentation (D or Dplus).}

\item{lasso_seg_model}{The model for the lasso matrix segmentation.}

\item{debug}{Logical, indicating whether to print debug information.}

\item{n_cores}{The number of cores to use for parallel processing.}

\item{make_symmetric}{Logical, indicating whether to make the matrix symmetric.}

\item{ds_fn}{The function to use for downsampling.}

\item{seg_algorithm}{The algorithm to use for segmentation.}
}
\value{
A list containing the processed and unprocessed matrices.
}
\description{
This function preprocesses a matrix by applying fast combining of , outlier removal, and fused 2D lasso filtering.
}
