% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickheat.R
\name{quickheat_gg}
\alias{quickheat_gg}
\title{Quick Heatmap Generation (ggplot)}
\usage{
quickheat_gg(mat)
}
\arguments{
\item{mat}{A matrix or data that can be converted into a matrix.}

\item{direction}{up or down}
}
\value{
A heatmap of the input data.
}
\description{
This function generates a heatmap from a given matrix using the ComplexHeatmap package. 
If the input is not a matrix, it will be converted to a matrix before generating the heatmap.
The heatmap does not include clustering or dendrograms, and does not display row or column names.
}
\examples{
mat <- matrix(rnorm(100), 10, 10)
quickheat(mat)

}
