% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_minmaxnorm.R
\name{scn_minmaxnorm}
\alias{scn_minmaxnorm}
\title{Perform SCN and Min-Max Normalization on a Count Matrix}
\usage{
scn_minmaxnorm(count_mat)
}
\arguments{
\item{count_mat}{A numeric matrix representing the count data to be normalized.}
}
\value{
A numeric matrix with the same dimensions as `count_mat`, where the data has been symmetrically
normalized and scaled using min-max normalization, with zero rows and columns preserved in their original
order.
}
\description{
This function applies symmetric normalization (SCN) and min-max normalization to a given count matrix.
It first makes the matrix symmetric if it is square, removes rows and columns with all zeros, applies SCN,
then applies min-max normalization to scale the values back to the original scale. It finally reintegrates
the zero rows and columns in their original positions and returns the normalized matrix with original row
and column names.
}
\examples{
# Generate a random count matrix
set.seed(123)
count_mat <- matrix(rnorm(100), nrow = 10)
# Apply SCN and min-max normalization
normalized_mat <- scn_minmaxnorm(count_mat)
}
\references{
Performs Sequential Component Normalization as described by Cournac. Coded using details in the manuscript. Cournac A, Marie-Nelly H, Marbouty M, Koszul R, Mozziconacci J. Normalization of a chromosomal contact map. BMC Genomics. 2012;13: 436. doi:10.1186/1471-2164-13-436
John C. Stansfield, Kellen G. Cresswell, Vladimir I. Vladimirov, Mikhail G. Dozmorov, HiCcompare: an R-package for joint normalization and comparison of HI-C datasets. BMC Bioinformatics. 2018 Jul 31;19(1):279. doi: 10.1186/s12859-018-2288-x.
}
