% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_lambda.R
\name{select_lambda}
\alias{select_lambda}
\title{Select Optimal Lambda Value}
\usage{
select_lambda(
  trt_ratio_mat,
  lambda = NULL,
  eps = 1e-04,
  maxIter = 10000,
  thresh = 0.1
)
}
\arguments{
\item{trt_ratio_mat}{A numeric matrix representing the treatment ratio matrix.}

\item{lambda}{A numeric vector of lambda values to test. If NULL (default), the function will use internally generated lambda values.}

\item{eps}{A numeric value specifying the convergence threshold for the fusedLattice function. Default is \code{1e-4}.}

\item{maxIter}{An integer specifying the maximum number of iterations for the fusedLattice function. Default is \code{1e4}.}

\item{thresh}{A numeric threshold value for filtering lasso values in plotting. Default is \code{0.1}.}
}
\value{
A numeric value representing the selected optimal lambda.
}
\description{
Selects the optimal lambda value for the fused Lasso regression using the fusedLattice method. It computes various statistics and generates plots to aid in selecting the lambda that best fits the data.
}
\details{
The function performs the following steps:
\itemize{
  \item Applies the fusedLattice function over a range of lambda values on the provided treatment ratio matrix.
  \item Calculates the number of non-zero coefficients in the beta estimates for each lambda.
  \item Computes lasso test values and summarizes them using statistical measures such as mean, median, standard deviation, kurtosis, and skewness.
  \item Generates various plots including density ridges, histograms, and line plots of statistical summaries against lambda.
  \item Determines inflection points in the mean lasso test values versus lambda curve using the \code{inflection} package.
  \item Selects the optimal lambda based on the inflection points, typically using the median of the lambda values corresponding to these points.
}
}
\examples{
\dontrun{
# Example usage:
library(glmgen)
library(pbapply)
library(dplyr)
library(ggplot2)
library(ggridges)
library(inflection)
library(e1071)
library(ggrepel)

# Create a treatment ratio matrix
trt_ratio_mat <- matrix(rnorm(100), nrow = 10)

# Select the optimal lambda
optimal_lambda <- select_lambda(trt_ratio_mat)
}
}
