% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_pooling_parallel.R
\name{sum_pooling_parallel}
\alias{sum_pooling_parallel}
\title{Sum Pooling with Parallel Processing Function}
\usage{
sum_pooling_parallel(mat, factor, ncores = parallel::detectCores())
}
\arguments{
\item{mat}{A numeric matrix to be pooled.}

\item{factor}{The pooling factor for aggregating matrix elements.}

\item{ncores}{The number of CPU cores to use for parallel processing (default is the number of detected cores).}
}
\value{
A pooled numeric matrix using sum pooling.
}
\description{
This function performs sum pooling on a matrix with parallel processing support.
}
\examples{
mat <- matrix(1:16, nrow = 4)
sum_pooling_parallel(mat, factor = 2)

}
