% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDistributions.R
\name{testDistributions}
\alias{testDistributions}
\title{Plot distribution AIC fit values}
\usage{
testDistributions(values)
}
\arguments{
\item{values}{A numeric vector containing the data for which the distribution fitting should be assessed.}
}
\value{
A ggplot object displaying a bar chart of AIC scores for each distribution tested, arranged from the lowest to the highest AIC.
}
\description{
This function evaluates and compares the goodness of fit for various statistical distributions
applied to a given dataset using the Akaike Information Criterion (AIC). It includes transformations
and several distribution fittings to identify the best model by AIC score.
}
\examples{
# Generating some random data
set.seed(123)
data <- rnorm(100)
# Testing distributions on the generated data
testDistributions(data)

}
