% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_read_correction_list.R
\name{total_read_correction_list}
\alias{total_read_correction_list}
\title{Perform Read Correction on a List of Matrices}
\usage{
total_read_correction_list(matrices)
}
\arguments{
\item{matrices}{A list of numeric matrices of the same dimensions. Each matrix
will be adjusted relative to the matrix with the smallest total sum.}
}
\value{
A list containing the adjusted matrices.
}
\description{
Adjusts the values in each matrix of the list (except for the one with the smallest
total read count) to match the total sum of the matrix with the smallest read count.
The adjustment is based on the ratio of total sums between each matrix and the one
with the smallest total sum. This function aims to equalize the total sums of all
matrices with that of the matrix with the smallest total sum.
}
\examples{
mat1 <- matrix(c(1, 2, 3, 4), nrow=2)
mat2 <- matrix(c(2, 4, 6, 8), nrow=2)
mat3 <- matrix(c(3, 6, 9, 12), nrow=2)
adjusted_matrices <- total_read_correction_list(list(mat1, mat2, mat3))
adjusted_matrices[[1]] # Adjusted mat1
adjusted_matrices[[2]] # Adjusted mat2
adjusted_matrices[[3]] # Original mat3 (smallest read count matrix in this case)
}
