% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb_parallel_compare.R
\name{zinb_parallel_compare}
\alias{zinb_parallel_compare}
\title{parallel_compare}
\usage{
zinb_parallel_compare(
  long_ctrl_trt_df,
  mc.cores = 1,
  interactive = F,
  thresh = 0,
  cap = Inf,
  fam = glmmTMB::truncated_nbinom2(),
  influence = F,
  return_models = F,
  offset = T,
  skip_zi = T,
  total_reads_ctrl = NULL,
  total_reads_trt = NULL,
  glm_fn = "glmmTMB",
  bias = F,
  bootstrap = F,
  n_straps = 1000,
  control_args = glmmTMB::glmmTMBControl(optimizer = optim, collect = F, profile = T,
    optArgs = list(method = "CG"), parallel = 1),
  single_model_cores = 1
)
}
\arguments{
\item{long_ctrl_trt_df}{A long-format data frame with control and treatment ratio data.}

\item{mc.cores}{The number of CPU cores to use for parallel processing (default is 1).}

\item{interactive}{A logical value indicating whether to plot the results interactively (default is FALSE).}

\item{thresh}{the hurdle value, diving the binomial (zero or not) and the zero-inflated negative binomial (count) models. Values below this are not used in count models.}

\item{cap}{The outlier threshold.}

\item{influence}{Should influence measures be calculated?}

\item{return_models}{Should the models be returned?}

\item{single_model_cores}{The number of cores to use for single models.}
}
\value{
A data frame with the treatment effect comparison results.
}
\description{
This function performs a parallel comparison of the treatment effect of a control and treatment group.
}
\examples{
dontrun{
parallel_compare(long_ctrl_trt_df, mc.cores = 2, interactive = TRUE) 
}
}
