% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb_parallel_compare_glm.R
\name{zinb_parallel_compare_glm}
\alias{zinb_parallel_compare_glm}
\title{parallel_compare}
\usage{
zinb_parallel_compare_glm(
  long_ctrl_trt_df,
  interactive = F,
  thresh = 0,
  cap = Inf,
  debug = F,
  plot = F,
  cores = 1,
  return_models = F
)
}
\arguments{
\item{long_ctrl_trt_df}{A long-format data frame with control and treatment ratio data.}

\item{interactive}{A logical value indicating whether to plot the results interactively (default is FALSE).}

\item{thresh}{the hurdle value, diving the binomial (zero or not) and the zero-inflated negative binomial (count) models. Values below this are not used in count models.}

\item{cap}{The outlier threshold.}

\item{debug}{A logical value indicating whether to enter the browser at the start of the function.}

\item{plot}{set to plot a heatmap of the results}

\item{cores}{The number of CPU cores to use for parallel processing (default is 1).}
}
\value{
A data frame with the treatment effect comparison results.
}
\description{
This function performs a parallel comparison of the treatment effect of a control and treatment group.
}
\examples{
dontrun{
parallel_compare(long_ctrl_trt_df, mc.cores = 2, interactive = TRUE) 
}
}
