% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zminmaxnorm.R
\name{zminmaxnorm}
\alias{zminmaxnorm}
\title{Z-Score Min-Max Normalization Function}
\usage{
zminmaxnorm(x)
}
\arguments{
\item{x}{A numeric vector to be normalized.}
}
\value{
A numeric vector with z-score min-max normalized values.
}
\description{
This function takes a numeric vector, standardizes it (z-score normalization),
and then performs min-max normalization on the standardized values, scaling them
to the range [0, 1].
}
\examples{
x <- c(1, 2, 3, 4, 5)
zminmaxnorm(x)

}
