% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zminmaxnorm.R
\name{ztrans}
\alias{ztrans}
\title{Z-Transformation Function}
\usage{
ztrans(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with z-transformed values.
}
\description{
This function takes a numeric vector and performs z-transformation.
}
\examples{
x <- c(1, 2, 3, 4, 5)
ztrans(x)

}
