% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dt_with_map_and_gc.R
\name{create_dt_with_map_and_gc}
\alias{create_dt_with_map_and_gc}
\title{Create Data Table with Mappability and GC Content}
\usage{
create_dt_with_map_and_gc(
  chr_val = "chr15",
  start_val = 61982445,
  end_val = 61988445,
  genome_val,
  map_fn = "/code/R/basepairC/data/mm10/k24.umap.bedgraph",
  mnase_model = system.file(package = "basepairC",
    "data/bias_models/mnase_mm10_kmer_map_6mer.txt"),
  mnase_model_8bp = system.file(package = "basepairC",
    "data/bias_models/mnase_mm10_kmer_map_8mer.txt"),
  bsgenome_obj = BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
)
}
\arguments{
\item{chr_val}{The chromosome for which to create the data table (default: "chr15").}

\item{start_val}{The start position of the genomic region (default: 61982445).}

\item{end_val}{The end position of the genomic region (default: 61988445).}

\item{genome_val}{The genome object (e.g., BSgenome object for the specified organism).}

\item{map_fn}{The file path to the mappability data (default: path to mm10 k24 umap bedgraph).}

\item{mnase_model}{The file path to the MNase bias model for 6-mers (default: path to mnase_mm10_kmer_map_6mer.txt in basepairC package).}

\item{mnase_model_8bp}{The file path to the MNase bias model for 8-mers (default: path to mnase_mm10_kmer_map_8mer.txt in basepairC package).}

\item{bsgenome_obj}{The BSgenome object for the specified organism (default: BSgenome.Mmusculus.UCSC.mm10).}
}
\value{
A data table with columns for chromosome, start position, end position, sequence, and calculated mappability and GC content for different window sizes.
}
\description{
This function creates a data table containing genomic regions specified by the user along with their mappability and GC content. It reads a mappability file for a given genome and calculates mappability and GC content for specified window sizes.
}
\examples{
dt_with_map_and_gc <- create_dt_with_map_and_gc(chr_val = "chr15", start_val = 61982445, end_val = 61988445, genome_val = BSgenome.Mmusculus.UCSC.mm10)

}
