% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnase_bias_6bp.R
\name{mnase_bias_6bp}
\alias{mnase_bias_6bp}
\title{Calculate MNase bias in a Sliding Window}
\usage{
mnase_bias_6bp(
  chromosome,
  position,
  genome = BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10,
  window_size = 6,
  n_cores = parallel::detectCores()/2,
  mnase_model = system.file(package = "basepairC", "data/mnase_model_6bp.rds")
)
}
\arguments{
\item{chromosome}{The chromosome for which to calculate MNase bias.}

\item{position}{A vector of positions on the chromosome to calculate MNase bias for.}

\item{genome}{The genome object from which to retrieve sequence data. By default, this is set to the mouse genome (BSgenome.Mmusculus.UCSC.mm10).}

\item{window_size}{The size of the sliding window for GC content calculation. Default is 5.}

\item{n_cores}{The number of cores to use for parallel processing. Default is half of the available cores detected on the system.}
}
\value{
A vector of GC content values for each position, calculated within the specified sliding window.
}
\description{
This function calculates the MNase bias within a sliding window across specified genomic positions on a chromosome.
}
\examples{
# Calculate GC content for positions 100 to 200 on chromosome 1
mnase_bias_test <- mnase_bias_win(
  chromosome = "chr1",
  position = 100:200,
  genome = BSgenome.Mmusculus.UCSC.mm10,
  window_size = 6
)

}
