% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebalance_observed.R
\name{rebalance_observed}
\alias{rebalance_observed}
\title{Rebalance Observed Samples}
\usage{
rebalance_observed(sample_mat, filter_type = "sorted")
}
\arguments{
\item{sample_mat}{A matrix or data frame containing the samples to be rebalanced. Numeric columns will be processed, while non-numeric columns will be kept intact.}

\item{filter_type}{A character string indicating the type of filtering to be applied. 
Can be either "sorted" (default, meaning to sort by the proportion of observed samples at each location followed by counts)
 or "ranked" (which takes the product of the two aforementioned factors and ranks them).}
}
\value{
A data table containing the rebalanced samples, with numeric columns equalized according to the selected filter type.
}
\description{
This function equalizes sample coverage by processing numeric columns in the input sample matrix.
It separates numeric and non-numeric columns, calculates total counts and non-zero counts, and 
rebalances the observed samples based on the specified filter type ("sorted" or "ranked").
}
\details{
The function first identifies numeric columns in the input matrix. It then separates numeric and non-numeric columns,
calculates the total counts and the total non-zero counts for the numeric columns, and determines the minimum non-zero count.
Based on the filter type, the function either sorts or ranks the data to filter out excess reads, ensuring that the resulting
data table has equalized sample coverage across all numeric columns.
}
\examples{
\dontrun{
sample_data <- data.table::data.table(A = c(10, 0, 5), B = c(3, 5, 0), C = c("Type1", "Type2", "Type1"))
rebalanced_data <- rebalance_observed(sample_data, filter_type = "sorted")
}

}
