% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb_parallel_compare_gam.R
\name{zinb_parallel_compare_gam}
\alias{zinb_parallel_compare_gam}
\title{parallel_compare}
\usage{
zinb_parallel_compare_gam(
  long_ctrl_trt_df,
  mc.cores = 1,
  interactive = F,
  thresh = 0,
  cap = Inf,
  fam = "nb",
  influence = F,
  return_models = F,
  offset = T,
  skip_zi = T,
  total_reads_ctrl = NULL,
  total_reads_trt = NULL,
  mnase = F,
  bootstrap = F,
  n_straps = 1000,
  mnase_mat = NULL,
  single_model_cores = 1,
  link = "log",
  extra_formula_terms = "",
  ridge = F,
  lasso1d_filter = F,
  restricted_mnase = T,
  base_formula = "value ~   treatment",
  robust_cov = F
)
}
\arguments{
\item{long_ctrl_trt_df}{A long-format data frame with control and treatment ratio data.}

\item{mc.cores}{The number of CPU cores to use for parallel processing (default is 1).}

\item{interactive}{A logical value indicating whether to plot the results interactively (default is FALSE).}

\item{thresh}{the hurdle value, diving the binomial (zero or not) and the zero-inflated negative binomial (count) models. Values below this are not used in count models.}

\item{cap}{The outlier threshold.}

\item{fam}{family object from the mgcv package.}

\item{influence}{Should influence measures be calculated?}

\item{return_models}{Should the models be returned?}

\item{offset}{A logical value indicating whether to include an offset term in the model.}

\item{skip_zi}{A logical value indicating whether to skip the zero-inflated model.}

\item{total_reads_ctrl}{The total number of reads in the control group.}

\item{total_reads_trt}{The total number of reads in the treatment group.}

\item{mnase}{A logical value indicating whether to include MNase bias correction.}

\item{bootstrap}{A logical value indicating whether to perform bootstrapping.}

\item{n_straps}{The number of bootstrap samples.}

\item{mnase_mat}{A numeric matrix representing the MNase bias matrix.}

\item{single_model_cores}{The number of cores to use for single models.}

\item{link}{The link function for the negative binomial model.}

\item{ridge}{Apply ridge penalty to treatment? Default FALSE (takes a bit longer).}

\item{lasso1d_filter}{Apply 1D LASSO filter to covariates? Default TRUE.}

\item{restricted_mnase}{Apply MNase bias correction only to the cell of interest? Default TRUE.
Changing this value to false models the entire matrix for MNase, which can be slow and memory-intensive.}
}
\value{
A data frame with the treatment effect comparison results.
}
\description{
This function performs a parallel comparison of the treatment effect of a control and treatment group.
}
\examples{
dontrun{
parallel_compare(long_ctrl_trt_df, mc.cores = 2, interactive = TRUE) 
}
}
